// types/job.ts
// types/job.ts
export interface Job {
    id: number;
    title: string;
    company: string;
    location: string;
    state: string;
    type: string;
    salary: string;
    department: string;
    posted: string;
    description: string;
    skills: string[];
    education: string;
    experience: string;
    industry: string;
    occupation: string;
    benefits: string[];
  }